pyproject.toml
[tool.poetry]
name = "realtime-mrs"
version = "0.1.0"
description = "Realtime EI visualization tool"
authors = ["User <user@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
psychopy = "^2025.1.1"
pygame = "^2.6.1"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
